saxs-ensemble-fit/gui/requirements.txt
Flask==0.10.1
Flask-Login==0.3.2
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.1
Flask-WTF==0.12
Flask-Admin==1.3.0
Flask_Script==2.0.5
Flask-Bootstrap==3.3.5.7
Jinja2==2.8
MarkupSafe==0.23
SQLAlchemy==1.0.9
WTForms==2.0.2
Werkzeug==0.10.4
flipflop==1.0
blinker==1.4
dominate==2.1.16
Tempita==0.5.2
itsdangerous==0.24
pbr==1.8.1
pytz==2015.7
sqlalchemy-migrate==0.10.0
sqlparse==0.1.18
